Avoid server hang on restarts
When restarting, the entrypoint was appending to existing configuration always, without checking if it was already filled. As such, since the 1st restart, frontend and backend names got duplicated and the server stopped working. In an automated restart environment, the file was only growing constantly.

Now we generate the full config and overwrite it always.